refactor(core): перенос config.py в подпакет core и добавление shim
• Полный код модуля скопирован в release_tool/core/config.py.
• core/__init__.py обновлён для импорта Config и load_config из нового расположения.
• Старый release_tool/config.py заменён лёгкой shim-обёрткой, сохраняющей
  обратную совместимость (реэкспорт + вызов main при запуске).
• План рефакторинга (realise_tool_refactornig.md) обновлён: шаг 1 помечен как выполненный.
