+++ core/__init__.py
@@ -8,7 +8,7 @@
 from __future__ import annotations
 
 # Переэкспорт существующих реализаций, пока они ещё находятся в корне
-from ..config import Config, load_config  # noqa: F401
+from .config import Config, load_config  # noqa: F401
 from ..git import GitRepo  # noqa: F401
 from ..packages import Package, iter_release_packages  # noqa: F401
